# [runtime] add count=1 special case for `implode()` function